// mobile/interfaces/home/realEstate.ts
export interface RealEstate {
  id_photo: number;
  id_real_estate_photo: number;
  id_real_estate: number;
  id_user: number;
  url: string;
  title: string;
  description: string;
  public_id: string;
  email: string;
  cellphone_number: string;
  available: boolean;
  amount_bedroom: number;
  price: number;
  amount_bathroom: number;
  square_meter: number;
  address: string;
}